requirements.txt
anyio==3.6.2
autopep8==2.0.2
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
fastapi==0.95.0
h11==0.14.0
hubspot-api-client==7.5.0
idna==3.4
pycodestyle==2.10.0
pydantic==1.10.7
python-dateutil==2.8.2
python-dotenv==1.0.0
requests==2.28.2
six==1.16.0
sniffio==1.3.0
starlette==0.26.1
tomli==2.0.1
typing_extensions==4.5.0
urllib3==1.26.15
uvicorn==0.21.1
